Restructuring what data is stored during the simulation.
Originally (and now in folder SU2xSU2_preproduction), I store all
simulated configurations in an array and perform the desired
measurements once the simulation has finished. For long runs and large
lattice sizes, this becomes very memory intensive, motivating to change
when measurements are made.
The now in the folder SU2xSU2 employed approach is to only store the
chain of observations for specific measurements during the simulation.
These are written to disk and can be used later to compute the ensemble
average and associated error.
While both implementations work with the same simulation logic, the new
approach should be used to collect production run data.